Add script that this project is inspired by
The script `submit_jobs_in_directory.sh` is meant to be run from a
cronjob. It automatically tops off a SLURM queue to a specified
number of jobs.

The intent of this code is to generalize this in multiple ways,
which will be documented in the README of the next commit.